Add handle updates util. Remove unused code.
